pyproject.toml
[tool.poetry]
name = "dataset"
version = "2.3.10"
description = ""
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.94.1"
fastapi-utils = "0.2.1"
gunicorn = "20.0.4"
uvloop = "0.17.0"
httptools = "0.5.0"
python-multipart = "0.0.5"
python-json-logger = "2.0.2"
aiofiles = "0.6.0"
psycopg2 = "2.9.9"
bids-validator = "1.7.3"
python-dotenv = "0.19.1"
opentelemetry-exporter-jaeger = "1.6.2"
opentelemetry-instrumentation = "0.30b1"
opentelemetry-instrumentation-httpx = "0.30b1"
opentelemetry-instrumentation-asyncpg = "0.30b1"
opentelemetry-instrumentation-sqlalchemy = "0.30b1"
opentelemetry-instrumentation-fastapi = "0.30b1"
aioredis = "2.0.1"
SQLAlchemy = "1.4.44"
asyncpg = "0.29.0"
greenlet = "1.1.2"
fastapi-pagination = "0.9.3"
async-lru = "1.0.3"
aiokafka = "0.7.2"
requests = "2.27.1"
pilot-platform-common = "0.3.1"
pydantic = "1.8.2"
uvicorn = "0.17.6"
fastavro = "1.9.4"

[tool.poetry.dev-dependencies]
pytest-env = "0.6.2"
flake8 = "4.0.1"
pytest-asyncio = "0.18.3"
async-asgi-testclient = "1.4.11"
testcontainers = "4.7.2"
pytest = "7.2.0"
pytest-cov = "3.0.0"
pytest-httpx = "0.21.2"
Faker = "13.16.0"
pytest-random-order = "1.0.4"
jq = "1.3.0"
alembic = "1.8.1"
redis = "^4.4.2"

[tool.pytest.ini_options]
addopts = "-v --random-order"
testpaths = "tests"
asyncio_mode = "auto"

[tool.coverage.run]
concurrency = ["thread", "greenlet"]

[tool.coverage.report]
exclude_also = [
    "raise NotImplementedError",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
